CLOUD-2240: Compilation fix
<==== Mercurial data ====>
Branch: cloud_dev

+++ cloud/backend/libs/nx_clusterdb_map/unit_tests/src/nx_clusterdb_map_ut.cpp
@@ -237,7 +237,7 @@ private:
 
     KeyValuePair* randomPair(size_t index = 0)
     {
-        NX_ASSERT(0 <= index && index <= m_randomPairs.size());
+        NX_ASSERT((size_t)0 <= index && index <= m_randomPairs.size());
 
         if (index == m_randomPairs.size())
             m_randomPairs.emplace_back(KeyValuePair());
